+++ src/ReBalanced.API/ReBalanced.API.csproj
@@ -8,39 +8,39 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Ardalis.GuardClauses" Version="4.0.0"/>
+        <PackageReference Include="Ardalis.GuardClauses" Version="4.0.1"/>
         <PackageReference Include="AutoMapper" Version="11.0.1"/>
         <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0"/>
-        <PackageReference Include="Azure.Extensions.AspNetCore.Configuration.Secrets" Version="1.2.1"/>
-        <PackageReference Include="Azure.Identity" Version="1.5.0"/>
-        <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.6.0"/>
-        <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.2.0"/>
+        <PackageReference Include="Azure.Extensions.AspNetCore.Configuration.Secrets" Version="1.2.2"/>
+        <PackageReference Include="Azure.Identity" Version="1.6.0"/>
+        <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.7.0"/>
+        <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.3.0"/>
         <PackageReference Include="MediatR" Version="10.0.1"/>
         <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1"/>
-        <PackageReference Include="Microsoft.AspNet.WebApi.Core" Version="5.2.7"/>
+        <PackageReference Include="Microsoft.AspNet.WebApi.Core" Version="5.2.8"/>
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.0.0"/>
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.0.0"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.2"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0.2">
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.4"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0.4">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.2"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.AzureKeyVault" Version="3.1.22"/>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.4"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.1"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration.AzureKeyVault" Version="3.1.24"/>
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0"/>
         <PackageReference Include="Serilog" Version="2.10.0"/>
         <PackageReference Include="Serilog.AspNetCore" Version="5.0.0"/>
         <PackageReference Include="Serilog.Extensions.Hosting" Version="4.2.0"/>
         <PackageReference Include="Serilog.Sinks.Console" Version="4.0.1"/>
         <PackageReference Include="Serilog.Sinks.Seq" Version="5.1.1"/>
-        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.2.3"/>
-        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.2.3"/>
+        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.3.0"/>
+        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.3.0"/>
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\ReBalanced.Application\ReBalanced.Application.csproj"/>
-        <ProjectReference Include="..\ReBalanced.Infastructure\ReBalanced.Infastructure.csproj"/>
+        <ProjectReference Include="..\ReBalanced.Infrastructure\ReBalanced.Infrastructure.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
