+++ src/ReBalanced.Application/Services/Interfaces/IPortfolioService.cs
@@ -1,5 +1,4 @@
 ﻿using ReBalanced.Domain.Aggregates.PortfolioAggregate;
-using ReBalanced.Domain.Entities;
 
 namespace ReBalanced.Application.Services.Interfaces;
