+++ src/ReBalanced.Domain/ReBalanced.Domain.csproj
@@ -7,9 +7,9 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Ardalis.GuardClauses" Version="4.0.0" />
-        <PackageReference Include="TimeZoneConverter" Version="5.0.0" />
-        <PackageReference Include="UnitsNet" Version="4.121.0" />
+        <PackageReference Include="Ardalis.GuardClauses" Version="4.0.0"/>
+        <PackageReference Include="TimeZoneConverter" Version="5.0.0"/>
+        <PackageReference Include="UnitsNet" Version="4.121.0"/>
     </ItemGroup>
 
 </Project>
