--- src/ReBalanced.Infastructure/MBoum/CoinQuote.cs
@@ -1,57 +1,0 @@
-﻿// Generated by https://app.quicktype.io/
-
-namespace ReBalanced.Infastructure.MBoum;
-
-public class CoinQuote
-    {
-        public Meta? Meta { get; set; }
-        public CoinData? Data { get; set; }
-    }
-
-    public class CoinData
-    {
-        public double Price { get; set; }
-        // public double PriceChangePercentage1H { get; set; }
-        // public double PriceChangePercentage24H { get; set; }
-        // public double PriceChangePercentage7D { get; set; }
-        // public double PriceChangePercentage30D { get; set; }
-        // public double PriceChangePercentage60D { get; set; }
-        // public double PriceChangePercentage90D { get; set; }
-        // public double MarketCap { get; set; }
-        // public double MarketCapChangePercentage24H { get; set; }
-        // public double FullyDilutedMarketCap { get; set; }
-        // public double FullyDilutedMarketCapChangePercentage24H { get; set; }
-        // public long CirculatingSupply { get; set; }
-        // public long TotalSupply { get; set; }
-        // public long MaxSupply { get; set; }
-        // public double MarketCapDominance { get; set; }
-        // public long Rank { get; set; }
-        // public double Roi { get; set; }
-        // public double Low24H { get; set; }
-        // public double High24H { get; set; }
-        // public double Low7D { get; set; }
-        // public double High7D { get; set; }
-        // public double Low30D { get; set; }
-        // public double High30D { get; set; }
-        // public double Low90D { get; set; }
-        // public double High90D { get; set; }
-        // public double Low52W { get; set; }
-        // public double High52W { get; set; }
-        // public double LowAllTime { get; set; }
-        // public double HighAllTime { get; set; }
-        // public double LowAllTimeChangePercentage { get; set; }
-        // public double HighAllTimeChangePercentage { get; set; }
-        // public DateTimeOffset LowAllTimeTimestamp { get; set; }
-        // public DateTimeOffset HighAllTimeTimestamp { get; set; }
-        // public double LowYesterday { get; set; }
-        // public double HighYesterday { get; set; }
-        // public double OpenYesterday { get; set; }
-        // public double CloseYesterday { get; set; }
-        // public double PriceChangePercentageYesterday { get; set; }
-        // public double VolumeYesterday { get; set; }
-        // public double Turnover { get; set; }
-        // public double YtdPriceChangePercentage { get; set; }
-        // public string? Status { get; set; }
-        // public double Volume { get; set; }
-        // public double VolumeChangePercentage24H { get; set; }
-    }
