--- src/ReBalanced.Infastructure/ReBalanced.Infastructure.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Ardalis.GuardClauses" Version="4.0.0" />
-        <PackageReference Include="GeoCoordinate.NetCore" Version="1.0.0.1" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.2" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0.2">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="6.0.2" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.2" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.0.2" />
-        <PackageReference Include="Refit" Version="6.3.2" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\ReBalanced.Domain\ReBalanced.Domain.csproj" />
-    </ItemGroup>
-
-</Project>
