+++ src/ReBalanced.Infrastructure/Context.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using ReBalanced.Domain.Aggregates.PortfolioAggregate;
-using ReBalanced.Domain.Entities;
-using ReBalanced.Domain.Entities.Aggregates;
 using ReBalanced.Infastructure.EntityConfigurations;
 
 namespace ReBalanced.Infastructure;
