+++ src/ReBalanced.Infrastructure/EntityConfigurations/HoldingEntityTypeConfiguration.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using ReBalanced.Domain.Aggregates.PortfolioAggregate;
-using ReBalanced.Domain.Entities;
 
 namespace ReBalanced.Infastructure.EntityConfigurations;
