+++ tests/ReBalanced.Infrastructure.Tests/ReBalanced.Infrastructure.Tests.csproj
@@ -37,7 +37,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\ReBalanced.Infastructure\ReBalanced.Infastructure.csproj" />
+      <ProjectReference Include="..\..\src\ReBalanced.Infrastructure\ReBalanced.Infrastructure.csproj" />
     </ItemGroup>
 
 </Project>
